Improve the duality gaps computation
I refactored the function that computes the conditional probability so
so that I can compute the individual duality gaps without redundant
computations.

I changed the criterion for updating the whole bacth of duality gaps. If
the non-uniformity is superior to 0.9, meaning that less than 10% of the
time is spent on a uniform drawing, than I allocate these 10% to the
update of the full dual_gaps table.