Update functions Marginals.is_density and is_consistent
I want to use them, so the first one should have a control over the
value of the sum. The other should not print anything.

Also rename variables in sdca.
Regularization parameter was too long. regu is more readable.
Also remove Csiba's sampling scheme because it is unfeasible in the
structured cas.