Move up the duality gap estimate that was false
I was estimating teh duality gap with KL between the updated marginals and the primal marginals. That was wrong ! It did not mathc the real duality gap and was necessarily closer to 0 as soon as I took steps close to 1.

Also rename the variables and remove the [i] indexing on the marginals array which was potentially slow.