Stabilize log_entropy and log_KL
I get no more assertion errors when I compute the entropy from the
marginals.

Also debug the line search. The root could be larger than 1 and the
evaluations made in find_root_decreasing were necessary to prevent a
ValueError in safe_newton.

To avoid redundant calculations, I now pass the value of the function on
the bounds as an argument to safe_newton.