Change a lot of things
Catch FloatingPointErrors in utils entropy and KL.

Change the way I compute LogProbabities.entropy and KL.
I comparing really two value and the value on the cliques should be
greater than the one on the separations.

Change the return value of SDCA. Objectives now include the duality
gap, the primal objective, the dual objective and the timing in this
order. It now returns 4 arrays instead of 5.

Correct the docstring of SDCA.

Modify the function of marginals_to_centroid. The difference with the
empirical expectation is now done in SDCA. That allows me to keep track
of the ground truth centroid which is useful to compute the primal
objective.

Wrap the operations I do after each epoch in one inner function:
full_batch_update.

Change the way I compute the first derivative of the line search
function. I now use KLs to be more numerically stable.

Change the ordering of the values returned in annex.

Guarantee that log probabilities returned by the oracle are smaller than
0 = log(1). I used to get some 1e-12 that could screw things up.

Move some values from file ocr to file constant and parse.