Add the empirical initialization for sdca
I use a convex combination of the empirical with the uniform to avoid
the infinite values of the entorpy's derviatives on the borders of the
simplex.

Also correct a bug present since I use tqdm. Make the loop start from 1
to avoid the double full_batch_update at the beginning of the algo.