Split the monitoring and the non-uniform sampling
I used to monitor the evolution of the duality gap at the same time
that I updated the sampling table. I now monitor the gap independently,
while subtracting the time spent on this from the iteration time.

The sampling table is updated only through the stochastic iterations.
To account for the staleness of the estimates, I recommend using a
mixed scheme, ie 0 < non-uniformity < 1.

Also add the option to not add the data in tensor board.