Make the line search work with Newton
The inequality abs(fdf) > abs(dxold) / 2 was in the wrong sense.
It pushed the line search to do only bisection.

I probably introduced this bug when I put the not to prevent the
propagation of NaN. To make the code more readable, I remove the not
and put the Newton step first in the code.

With a precision 10-4, the algo now uses 3 steps on average,
without including the border evaluation. I can actually remove the
evaluation in 0, although it happens for a small cost.