Change emission features to a dense matrix
The sparse matrix made no sense because at the end of the day,
the parameter is made from all the words and is basically dense.

I also change the way to add the dictionary by using the numpy.unique
method. It is much faster than adding the vectors of size 75k at each
iteration.