Update new marginals again after safe newton
Safe newton does not return the result for the last function evaluation
but for the one before last. That led to a terrible bug : I was not
updating my dual variable as much as my primal variable. This is a bad
side effect of hiding the update of the marginals into the line search.
I need to correct that.