Finish the self-contained class Weights
Use the argument is_features_sparse to handle the different kinds of
inputs. This is suboptimal because this is not a real feature of the
weights but more of the data. But this is acceptable because the data
and the weights are inextricably linked.

Also remove the useless function label_scores